Add Progressing translate and add new wrong message

+++ src/main/kotlin/com/io/cache/TaskCache.kt
@@ -8,9 +8,9 @@ import com.io.util.GetBooleanViaT
 
 interface TaskCache {
 
-    suspend fun saveTask(chatId: String, messageId: Long, taskId: Long, state: LessonState)
-    suspend fun deleteMessageTask(messageId: Long, term: GetBooleanViaT<MessageToTask>)
+    suspend fun saveTask(chatId: String, messageId: Int, taskId: Long, state: LessonState)
+    suspend fun deleteMessageTask(messageId: Int, term: GetBooleanViaT<MessageToTask>)
 
     suspend fun getRandomTaskFromSection(sectionId: String, state: LessonState): Task?
-    suspend fun getCurrentTask(messageId: Long): Task
+    suspend fun getCurrentTask(messageId: Int): Task
 }
