+++ src/main/kotlin/com/io/cache/entity/MessageEntity.kt
@@ -4,20 +4,20 @@ import com.io.model.LessonState
 import com.io.model.MessageGroup
 
 data class MessageEntity(
-    val id: Long,
+    val id: Int,
     val chatId: String,
     val group: MessageGroup,
 )
 
 data class MessageToSection(
     val chatId: String,
-    val messageId: Long,
+    val messageId: Int,
     val sectionId: String
 )
 
 data class MessageToTask(
     val chatId: String,
-    val messageId: Long,
+    val messageId: Int,
     val taskId: Long,
     val lessonState: LessonState
 )
