+++ src/main/kotlin/com/io/util/TypeAliasFunc.kt
@@ -5,7 +5,7 @@ import com.io.cache.entity.SectionRuleEntity
 import com.io.cache.entity.UserEntity
 import com.io.model.MessageGroup
 
-typealias GetMessageEntityViaIntToMessageGroup = suspend (messageId: Long, group: MessageGroup) -> MessageEntity?
+typealias GetMessageEntityViaIntToMessageGroup = suspend (messageId: Int, group: MessageGroup) -> MessageEntity?
 typealias GetUserEntity = suspend () -> UserEntity?
 typealias GetBooleanViaT<T> = (T) -> Boolean
